ct-superRes/requirements.txt
torch
torchvision
numpy
matplotlib
Pillow
scikit-image
tqdm
pydicom
requests
